refactor(schemas): migrate Phase 27 endpoint schemas — 2 files, 10 classes + 4 enums (#6042)
- analytics_quality.py: migrate QualityGrade, MetricCategory enums + QualityMetric,
  HealthScore, PatternDistribution, ComplexityMetrics, QualitySnapshot
  → schemas_analytics.py
- analytics_architecture.py: migrate PatternType, ConsistencyLevel enums +
  PatternMatch, PatternConsistency, ArchitectureLayer, ArchitectureReport,
  AnalysisRequest→ArchitectureAnalysisRequest → schemas_analytics.py

Renames to avoid generic-name collisions:
- AnalysisRequest → ArchitectureAnalysisRequest
